resumescren/backend/requirements.txt
flask
flask-cors
python-docx
pdfminer.six
scikit-learn
numpy
